pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Max Burzer <maximilian.burzer@student.kit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
tk = "^0.1.0"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
opencv-python = "^4.9.0.80"
types-pillow = "^10.1.0.20240106"
torch = "^2.1.2"
pandas-stubs = "^2.1.4.231227"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
